feat: move theme context to zustand store
