+++ src/app/components/command/command-menu.tsx
@@ -16,11 +16,11 @@ import {
 } from '@/app/components/ui/command'
 import { useApp } from '@/app/contexts/app-context'
 import { usePlaylists } from '@/app/contexts/playlists-context'
-import { useTheme } from '@/app/contexts/theme-context'
 import { useSongList } from '@/app/hooks/use-song-list'
 import { ROUTES } from '@/routes/routesList'
 import { subsonic } from '@/service/subsonic'
 import { usePlayerActions } from '@/store/player.store'
+import { useTheme } from '@/store/theme.store'
 import { Albums } from '@/types/responses/album'
 import { ISimilarArtist } from '@/types/responses/artist'
 import { ISong } from '@/types/responses/song'
