+++ src/app/observers/theme-observer.tsx
@@ -1,0 +1,26 @@
+import { useEffect } from 'react'
+import { useTheme } from '@/store/theme.store'
+
+export function ThemeObserver() {
+  const { theme } = useTheme()
+
+  useEffect(() => {
+    const root = window.document.documentElement
+
+    root.classList.remove('light', 'dark')
+
+    if (theme === 'system') {
+      const systemTheme = window.matchMedia('(prefers-color-scheme: dark)')
+        .matches
+        ? 'dark'
+        : 'light'
+
+      root.classList.add(systemTheme)
+      return
+    }
+
+    root.classList.add(theme)
+  }, [theme])
+
+  return null
+}
