+++ src/store/theme.store.ts
@@ -1,0 +1,34 @@
+import { merge } from 'lodash'
+import { devtools, persist, subscribeWithSelector } from 'zustand/middleware'
+import { immer } from 'zustand/middleware/immer'
+import { createWithEqualityFn } from 'zustand/traditional'
+import { IThemeContext, Theme } from '@/types/themeContext'
+
+export const useThemeStore = createWithEqualityFn<IThemeContext>()(
+  subscribeWithSelector(
+    persist(
+      devtools(
+        immer((set) => ({
+          theme: 'system',
+          setTheme: (theme: Theme) => {
+            set((state) => {
+              state.theme = theme
+            })
+          },
+        })),
+        {
+          name: 'theme_store',
+        },
+      ),
+      {
+        name: 'theme_store',
+        version: 1,
+        merge: (persistedState, currentState) => {
+          return merge(currentState, persistedState)
+        },
+      },
+    ),
+  ),
+)
+
+export const useTheme = () => useThemeStore((state) => state)
