+++ src/types/themeContext.ts
@@ -1,0 +1,6 @@
+export type Theme = 'dark' | 'light' | 'system'
+
+export interface IThemeContext {
+  theme: Theme
+  setTheme: (theme: Theme) => void
+}
